[roll] Roll fuchsia [bt][a2dp] CodecNegotiation prefers a direction
Now that we are presented with the opportunity for connecting both
source and sink endpoints, we need to choose which direction to connect
when we have both.

Add a preferred direction to the CodecNegotiation, and prefer the
correct direction in all of the A2DP profiles. We currently prefer
to be Source in bt-a2dp.

Original-Bug: 61530
Test: fx test bt-a2dp-lib-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/455495
Original-Revision: 3e6c2140de06bcef0e3552740acaa829abb3d681
GitOrigin-RevId: 4932090812512fd91c36d40e6a94bafdce151c74
Change-Id: I894604f26dca75880b24f89cb426d16c57ef9fee