[roll] Roll fuchsia [debugger] Add missing C++ operators.
Fills out the operator enum sufficiently to allow all "operator X"
definitions to be parsed. Most of the operators were in-place update
operators like "++" and "+=". These are added but unimplemented and will
give an error message. They are uncommon to use in the debugger.

The main missing operator was ~ which required some special handling to
disambiguate from destructor names. Becuase the debugger doesn't always
have the correct context and a better implementation seemed
unnecessarily complex for this uncommon operator, ~ followed by a word
is always treated as a name. To express a binary not, the caller needs
to add a space or another token like a paren.

Original-Bug: 49698

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/458454
Original-Revision: 3a720a5ad0fcfa71b37ad0748cf0739a9c77f860
GitOrigin-RevId: 10b395e364f281fd0188f19c947d4a340ed9b495
Change-Id: I2692f90ba4c536f48d755860d9382bb20ee659fd