[roll] Roll infra/recipes [checkout] Use consistent return type for CheckoutResults.changed_files
See go/bbid/8861544249433789728 of an example of a static-checks build
which uses a recipes.git GerritChange, and therefore has no changed
files. This should not be a recipe crash: there are just no files to
check, and we should return green.

This change swaps the `None` return values to `[]` for
CheckoutResults.changed_files so that users don't have to check
explicitly for None before iterating over the list.

Original-Bug: 58899
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/458915
Original-Revision: ce00ef665d9cc15a0b8601db41950c1ee48585af
GitOrigin-RevId: e7371507ab38cb1e81e768c3176689b2898a0b94
Change-Id: I46a3fa0527d79f9fff3d894409443d7b693e59ad