[roll] Roll fuchsia [session] Implement add command in session_control
session_control add <element spec> should add an element to the
session, if the session exposes the ElementManager protocol.

Ideally the session manager would expose the session's element manager
directly in the session manager cml.

If this is not possible, another way of doing this is to have the
session_manager component expose the ElementManager, just like it
exposes fuchsia.session.Launcher, and proxy incoming requests to the
session.

Can be closed when session_control can be used to launch an element in a session.

Original-Bug: 47125
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/436207
Original-Revision: 0727d493af31e22fd5db6e9ed223db4049b68d6d
GitOrigin-RevId: 882c743e0b8fe95862b768ea89ffaf6156d5daf5
Change-Id: Iab3d2b5d7f6ff91de535e7b2bf1e56ca5b4901e5