[roll] Roll fuchsia [component_manager] Update Security Policy Runtime Config.
This change flattens the structure of the CapabilityAllowlistEntries.
This has been changed primarily so that capability paths can be
correctly expressed in the context of their "source" either Self or
Framework. Without this change the system is inflexible as its difficult
to express a Directory that is sourced from the "<component_manager>".
Explicit definition of the source while slightly more verbose increases
the actual readability of the policies as you can correct know exactly
what it is targetting without additional context.

This configuration file will be kept small anyway and is an internal
security feature without a public API surface.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/458340
Original-Revision: 0e27978bc1bfb852abd3bfee788d67bfda9768fe
GitOrigin-RevId: 21285017104a5d59575f4e0ea7e99f0525134b33
Change-Id: Ided8ea29af95b6aeeceb9ccd7d6714e3596cb6d5