[roll] Roll fuchsia [kernel][x86] Multi-page support in MappingCursor
This extends the MappingCursor implementation to support multiple
physically discontiguous pages instead of just a single contiguous one.
In supporting this the MappingCursor class is extended to be more
object like and move all its data to be private with accessors.

The additional paddr support is not taken advantage of in this commit,
and this change should be perfectly semantic preserving.

Test: Covered by existing kernel unit tests.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/458265
Original-Revision: 0135d5f531731a0ad17040bbd4aa49610c65bf8d
GitOrigin-RevId: d7c32cb94de9643dcefbf32772e5aa68226d7984
Change-Id: I66389cecb8d6787e541197e5acfe865b0864fe72