[roll] Roll infra/recipes [fuchsia] Show "testing skipped" in build summary
Currently you need to go into a CQ build's results page to determine
whether it's green because all tests passed, or because the build graph was
unaffected and the recipe skipped running tests.

Adding this message to the summary markdown will make it visible from
builder pages, which will be especially useful when looking through the
history of a tryjob searching for the beginning and end of breakages.

Also add a period to the end of the text for consistency with summary
markdown messages from other recipes.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/459257
Original-Revision: d292801773e7eb6a87222288978bc2c26c7305ee
GitOrigin-RevId: 8061fb92f084dde9b2b25ae1958d890db12f29f8
Change-Id: I36cbce16a72f8b9d35163fd6c4b50376f0103543