[roll] Roll fuchsia [modular][flake] Fix race in ElementManagerTest watcher registration
The ProposeElementCreatesStory and ClosingElementControllerDeletesStory
tests use a StoryProviderWatcher to tell when the story is created
or deleted. The watcher does not get registered until the GetStories2
call (in SimpleStoryProvider.Watch) returns.

This ensures the watcher registration does not race with the story
being created in ProposeElement.

As a bonus, also assert that no stories have been created prior to
proposing, so all events are for the element story.

Original-Fixed: 65857
Test: fx test element_manager_test -o -- \
    --gtest_filter=ElementManagerTest.ProposeStartsStory:ElementManagerTest.ClosingElementControllerDeletesStory \
    --gtest_repeat=1000 --gtest_break_on_failure=1
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/459003
Original-Revision: 43799051808094534a8110401fef4a45abc1d59f
GitOrigin-RevId: c4f15a1fe1e36e8f1ab6f59bb8220a470b6e97a5
Change-Id: I40dea1c66ebb3c3c4fcdbee8104afaa427991912