[roll] Roll fuchsia [wlan] Consolidate IEs in BssDescription
Two main code paths that this impacts:
1. The scan code paths as BssDescriptions are relayed from driver to the
   upper layer
2. The join request code path as BssDescription is included. Driver now
   has to parse the SSID from the IEs block.

In the scan code paths, SME can no longer merge the individual fields
parsed from beacons and probe responses. It now uses the IesMerger
introduced in a previous commit to merge these IEs.

On scan end, SME converts the FIDL BssDescription to a parsed Rust
BssDescription. If the parse fails, that result is discarded and a
counter in Inspect is incremented. The parsed Rust BssDescription is
very close to the older FIDL BssDescription, and is used everywhere else
in SME.

Original-Bug: fxbug.dev/63004
Test:
1. CQ
2. Ran ACTs tests:
  a. WlanPhyCompliance11ABGTest
  b. WlanScanTest

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/442678
Original-Revision: 18a927b90622363d0fe26642f7ed9cdd40dadc91
GitOrigin-RevId: 0fa426946609805a6c2762a6921d0e07c376608a
Change-Id: I8448cd8dbb342235d7ddd04047ec12eda19f6318