[roll] Roll topaz [build] Specify path to vendored python binary in exec_script
This updates the two exec_script() invocations in this repository to
specify the full path to the vendored python binary to use instead of
relying on the PATH or global script_executable to find it. This also
means that the path to the script, along with any other input files,
needs to be passed to exec_script()'s file_dependencies parameter so
that the generated ninja file re-generates when these inputs change.

Original-Bug: 63458
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/topaz/+/446495
Original-Revision: 6ecb8a44b8f307da9ac323f78039dbc6b696e768
GitOrigin-RevId: fbb11b7aa2779613f0037d8961f214d050a84120
Change-Id: Id4697d4e711ce32b6e3b8525ee271646ddaa073e