[roll] Roll fuchsia [omaha-client-integration-tests] use isolated system updater
This modifies the omaha-client-integration-tests so that it spins up a
system-updater using the the isolated cmx, rather than the one on
target (non-isolated).

This useful because otherwise, we'd have to OTA before running the test
with the most recent system-updater.

SECURITY: We need to modify the allowlists because we now have a new
system-updater cmx. This has the same security considerations as the
other integration tests that are on the allowlist, so it should be ok.

Test: fx test -o omaha-client-integration-tests
Original-Bug: 64602
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/458901
Original-Revision: 47e202cfff67ebb200a51fd282bc3f3c8d09aa88
GitOrigin-RevId: 746e741b580b2a946edcee49da1c25d9d3d25a6b
Change-Id: Ie8c22bbc161c8fb6bfa65bcfc3430d393ad5a03e