[roll] Roll fuchsia [omaha-client] allow rebooting to clear NO_SPACE
The final CL in the sequence for fxbug.dev/65492.

If the observer has set previous_out_of_space_failure in the
FidlServer and we get a call to `PerformPendingReboot`, we should
immediately reboot. This allows us to clear a condition in which a
previous update resolved a bunch of packages and ran out of space, but
we can't clear the packages with a garbage collection because the
packages are still in pkgfs' dynamic index.

This solution exists as a short term fix. The "more correct" approach
is coming when pkg-cache has a concept of retained packages (see
fxbug.dev/60169).

Original-Fixed: 65498
Test: add a unit test for PerformPendingReboot, fix previous unit
tests for that function
Test: add an integration test for the whole flow from pkg-resolver's
NO_SPACE to rebooting

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/458777
Original-Revision: 70fb1612ab2df183d97d03ef1dc45c56adaef42f
GitOrigin-RevId: cdadd114ab41236d9b8e591abc67a7ba0574ec5c
Change-Id: I60aaf276f7985cd4d3b6e5148ac458dd4a0565c8