[roll] Roll fuchsia [minfs] Enable LSan for tests
LSan exception was added for inspector-test, which is actually in
zircon/system/utest/inspector/. So remove the exception.

This unearthed a few leaks in several tests, mostly related to death
situations. Most of the leaks were fixed by moving code outside of the
death lambda. In on situation though (TakeBlockReservationsWithInodeReservationDies)
we were losing ownership of the transaction in the death clause. I
didn't have a good way to fix this, so I ended up keeping track of
the raw pointer to delete it afterwards.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/458937
Original-Revision: 32b39314b0cd2cfac32504739e3a71fe3878ddf1
GitOrigin-RevId: 3d450fd20a215ef6f42cd0563bfa45d5c386b78b
Change-Id: I0667c83980119a78524aef6764fb8855335ff010