[roll] Roll infra/recipes [buildbucket_util] Don't split unicode chars in build summary
The old summary truncation logic would occasionally split a line in the
middle of a multi-byte unicode character, which would produce a summary
that was not valid UTF-8. This causes infra failures when trying to
write the resulting text to the build proto's summary_markdown. For
example: https://ci.chromium.org/b/8861422871643300480

To avoid this, use a slightly smarter truncation strategy: if the split
would occur in the middle of a unicode character, we'll instead split
immediately before that character.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/459734
Original-Revision: d3671791797bc564981f5bc2e429d172055b1651
GitOrigin-RevId: 7e4b6ef0c77ed11a5d219d6c38f0b8796617a443
Change-Id: Ie9c0bec8de1fb24838945859acfeca3f5b443a5c