[roll] Roll infra/recipes [jiri] Turn red if patch_project is not in checkout
If patching fails because the specified project is not in the checkout,
it's likely because someone ran a builder in on a CL in a project that's
not included in the builder's checkout. This is not an infra error.

The latest version of `jiri patch` returns retcode 23 if this happens,
so we can use that to disambiguate between missing projects and other
patch errors.

Original-Fixed: 66079
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/459256
Original-Revision: 821b2e8973172e1b4b2008ec15c7b76ef62f8268
GitOrigin-RevId: 86ddcf1d4c71cb1faa30cfea005ad932c8d037bd
Change-Id: I087fc29232765a3585d1735f393c13d49991c832