[roll] Roll fuchsia [kernel][init] Don't print until after global ctors.
A fair amount of the printf machinery is now using C++, and relying on
the fact that global/static classes used in the stack have trivial
constructors.  For the most part, this is working out (although it
could, in theory, break at any time), but there is at least one edge
case which is failing.

If ENABLE_KERNEL_LL_DEBUG is defined, it will initialize the global
dlog_bypass_ boolean to true, telling the system to go directly out
the serial port instead of going to the debuglog's internal buffer.
During the first printf in top/main.cc (the "printing enabled"
banner), we end up calling...

1) dprintf
2) printf
3) vprintf
4) vfprintf
5) File::Write (thunks to stdout_'s lambda)
6) stdout_write_buffered
7) Linebuffer::Write
8) stdout_write
9) console_write

The vfprintf call (#4) takes a global FILE instance (stdout_), but
this is currently "ok" as this instance is successfully
data-segment-only initialized (currently, at least).  The trouble
starts when we get to console_write, which grabs a spinlock, and then
attempts to enumerate the callbacks stored in the global
print_callbacks list.

The print_callbacks list is a fbl DLL whose head pointer is still 0,
it has not been updated to the sentinel pointer value yet.  This trips
an assert very early on in boot which leads to a never ending double
fault on x64.

So, for now, make sure that global ctors are finished before we
perform our very first printf.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/459506
Original-Revision: 61fdef500de30f2b9a9e746418af7f4a07505e5a
GitOrigin-RevId: 3366bfc45c27c6189e06667b551c230997e57a98
Change-Id: I920f4f33ac502df068736dc5f88c8be1c79a99db