[roll] Roll fuchsia [SetUI][MessageHub] Roles Introduction
Roles are a way to group messengers together under
a label (or role signature) so that messages can be
targeted towards the group as a whole. This provides
a way to target messengers in a way that is broader
than addresses and more specific than a braodcast.

Roles may be specified at compile time as part of the
MessageHub definition and/or at run time with
generated, anonymous roles.

Original-Fixed: 64977
Test: fx test -o setui_service_tests
Multiply: setui_service_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/453884
Original-Revision: ed942e3fff809cfa4015eb6271f150d68fad1fbb
GitOrigin-RevId: dea92ac2eb1169f55540de0bbbe5d6402e499243
Change-Id: I7bd1bb0e61a3d2d3e5d88815aa6a0c0cc7d3d74d