[roll] Roll fuchsia [pkg-resolver] Raise blob fetch concurrency to 2
Investigation over both HTTP/1.1 and HTTP/2 showed that increasing
concurrency from 1 to 2 provides the vast majority of the benefit of
increasing concurrency with blob fetches, while limiting the maximum
memory overhead that blobfs can impose on the system by keeping blobs in
memory until they're written to the underlying storage.

This change also annotates the choice of 2 to that effect.

The `dedup_concurrent_blob_fetches` test is not re-enabled as it
requires a blob fetch concurrency of at least 4 to function. However,
this test separately validates the behavior that the concurrent queue
context merging tests also validate, so we're still protected against
regression by existing tests. Because we can't guarantee an ordering
for blob requests, additional testing for this would require an ability
to make blob fetches deterministic, which might make this change take
significantly longer than desired to complete.

fxbug.dev/65855 was filed to track this test.

Original-Fixed: 61394
Test: unignored pkg_resolver concurrency tests still succeed
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/458139
Original-Revision: afe4c6dc10f44ff3a80f5adcf971039681cd8fa9
GitOrigin-RevId: 405239396f88c5eef419365aa04ebb5fabb6c9ba
Change-Id: I3ce0d5f1e1ee6323de204c2b554d4f6e97eb3e53