[roll] Roll fuchsia [ffx][daemon] daemon.quit() safety check.
This adds a safety check to make sure not to run exit(0) inside unit
tests. This also removes unit tests that are covered in self-test which
invoke Quit as this can cause tests to exit prematurely and appear to
succeed even in the event of a failure.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/459201
Original-Revision: d316070227c48dd674a42380fef7c3dea156ce04
GitOrigin-RevId: fa065106bb3078abd6c6a34ad9de2d6baefeb136
Change-Id: Ie15eab851e1c570cdc291306986cd60123d46a83