[roll] Roll infra/recipes [auto_roller] Add a no_tryjobs option
If set, this will cause a roller to CQ+2 roll CLs, but CQ will submit
the CL without running any tryjobs.

The advantage of this over `force_submit` is that `no_tryjobs` rollers
will still respect the tree status, so they won't land roll CLs during
tree closures. This is useful for a particular repository that doesn't
have CQ enabled, and is pinned in an integration.git manifest for
convenience but not included in a Fuchsia checkout. Rolls of this repo
are essentially no-ops, so it's safe to not run any tryjobs on them, but
we don't want them polluting the CI console and running unnecessary CI
builds during tree closures.

Original-Bug: 66170
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/460030
Original-Revision: f15b5735bc4ece7abae92c845035e1f1c12c45aa
GitOrigin-RevId: 6f4f99da9a9143083959ff3e7414e237ead78d76
Change-Id: I8339e6b12a99373804f54217c80c502ccd0eded0