[roll] Roll fuchsia [kernel][vm] Decommit VMO on pager detach
When a pager-backed VMO was detached from the pager, the default
faulting behavior for its clones was to serve zero pages if a page was
not found in the parent. This sometimes caused applications executing
from blobs to fail in confusing ways after blobfs had shut down,
depending on how the zero pages were interpreted.

This commit changes the default and instead of returning a zero page for
the clone, propagates the ZX_ERR_NOT_FOUND error code through the fault
resolution path. It also decommits *all* committed pages from the parent
VMO (the one directly backed by the pager source). Decommitting all the
pages ensures that the failure model is consistent across all cases, and
does not depend on what pages happen to be committed in the parent at
the time.

Accessing pages in the parent or the clone after the pager source has
been detached will now consistently result in a fatal page fault (if
accessed via a mapping). A syscall on the VMO which requires pages to be
present will fail.

Note that this change only pertains to pages in the clone that are owned
by the parent, not pages that have been COW forked by the clone. All
forked pages will continue remaining valid since they are owned by the
clone itself.

Test: core-pager

Original-Bug: 54120
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/459078
Original-Revision: 579b625b87352803af0a4351a2a7d3feb1b02736
GitOrigin-RevId: 7bbf222d982a6399cb64a1186a8618b0cf89834e
Change-Id: Ie378a58d052cf7bb058b7c8ebba09dda26047a2c