[roll] Roll fuchsia [storage] Wire up the FVM partition limit to fshost.
fshost currently has some configurable partition limits for blobfs and
minfs (not currently set so this code path is still unused). And FVM
supports setting limits on partition sizes.

This patch wires up the values in fshost to FVM when a device is
detected.

Separates out the mock block devices from the block-watcher-test to a
separate header so these can be shared with the new block-watcher-tests.
There is no change to these with the exception of adding a field for
the max size set.

Original-Bug: 64055

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/452517
Original-Revision: ea1290f4cf906e72e59cb7b755c9267b027e6bed
GitOrigin-RevId: 9403812d54c319f62828fc256b2880de776ab08f
Change-Id: Idba5c55d59f63d2e6ffa3f404a82fe30d0493bfc