[roll] Roll fuchsia [component_manager] Add capability routing checking to GlobalPolicyChecker.
This change implements a new member function into the GlobalPolicyCheckr
that given a `CapabilitySource` and a target `Realm` returns Ok() if no
policies are violated else a PolicyError. This change does not integrate
the global policy checker into the capability routing path that will be
enabled in a subsequent patchset.

Original-Bug: 61864
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/459099
Original-Revision: 4b2d040c05691bf7f639ed4ce2071515ae932898
GitOrigin-RevId: 38aa4ab3f3f1e74b6e285119876e529b0911daa9
Change-Id: Ia43866860f1209c12daefc198393326d998d369c