[roll] Roll fuchsia [kernel][arm64][timer] add a unit test for the arm event stream timer
Validate that the event stream seems to be generating events at the rate
the cpu says it will. Err on the side of a false negative, since
anything in the system can generate events outside of the event stream,
even across cpus. Only look to see that the time spent in WFE is not too
long. Only enabled if the feature is turned.

NOTE: Running on qemu, kvm or otherwise, seems to be meaningless since
WFE either vmexits or falls through. The test will not fail, but the
measured values are extremely low, as if the instruction was largely a
nop. Potentially the same thing if running under any other hypervisor.
It may be worth simply disabling the event stream feature if the kernel
has detected its on a hypervisor, but there's not (currently) a good way
to universally detect this.

Original-Bug: 56818
Tested: run on real hardware, validate that the observed times are
reasonable and within the computed limits.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/458322
Original-Revision: 38e94994ab39993e71aba5676de80fc3f8cd0d55
GitOrigin-RevId: 9f0fbd9944d02866a52f7fd07eb3b99e384d336c
Change-Id: I0e5e6f886d9b080a31d50c62f3bf805330321c68