[roll] Roll fuchsia [logs] Refactor the structured log encoder into its own header file
Expose the structured log encoder externally by creating a
header-only library for the encoder that is part of the
structured logging backend. This will eventually be used
in macros.h to replace the encoding that currently
takes place in the backend.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/440534
Original-Revision: 655d09d51638d8d348dfc0cfeb9c59f11ead0dcb
GitOrigin-RevId: 60d2f58c5ca50aca49a038586cdb2f5be377b723
Change-Id: I3971d8b458a293c57e9303cc0409e22fbe8f00b4