[roll] Roll fuchsia [time][docs] Add initial documentation structure for time.
This creates the full structure but content on many of the child pages
is not yet populated in the interest of keeping small CLs and easy
review. The remaining content will be populated by Satsuki and I over
the next 1-2 weeks.

For convenience, rendered HTML is available at
https://fuchsia.devsite.corp.google.com/preview/jsankey/

Original-Bug: 65727
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/458658
Original-Revision: 1f6422c1e2cab69f21cfe318e2bbfaecfbdb1385
GitOrigin-RevId: 880931cf9b97d247d09225b7e1d7066b2f223acc
Change-Id: I7d73c3f3d4847df93e60a173fa9b46585514f3f1