[roll] Roll fuchsia [build][asan] Use -asan-use-private-alias=1
A change in how dependencies are propagated by GN through Rust libraries
causes:

[343.591] 284345.284347> ERROR: AddressSanitizer: odr-violation (0x23546df33ce0):
 [343.633] 284345.284347>   [1] size=29 '{{{symbol:_ZTSN3fsl13SocketDrainer6ClientE}}}' ../../src/lib/fsl/socket/socket_drainer.cc
 [343.633] 284345.284347>   [2] size=29 '{{{symbol:_ZTSN3fsl13SocketDrainer6ClientE}}}' ../../src/lib/fsl/socket/socket_drainer.cc

which has been causing the GN roll to fail for a while now.

That symbol is internal to the compiler, demangles to "typeinfo name for
fsl::SocketDrainer::Client".

Making typeinfo private to the module seems reasonable (?), which is I
believe what this clang flag does (and what gcc does by default). It
appears this is the plan for clang to default to in
https://reviews.llvm.org/D92078, and this appears to resolve the ASAN
ODR violation in local testing.

Original-Bug: 66129
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/460659
Original-Revision: 25583e3f396c0ea827b6a96a1d62aeda85453b29
GitOrigin-RevId: 7746627ee18ec40ba5d7df92844513d909f12a69
Change-Id: I16d6a2d1f4e592f3590556bd1b4d2e64ee088a9b