[roll] Roll fuchsia [stdcompat] Add version header.
Provide a <version> polifyll with 'string_view' feature test symbol,
so the header is introduced in a different cl.

The goal is whenever a polyfill is added, the set of feature test macros
that are needed, are either provided by the platform or defined in this
header.

Added test that verifies for std14,17 and 20 that the symbol is either
undefined(cpp14), or at a specific draft version(17, 20) depending on
the platform cplusplus version.

Original-Bug: 57356 #comment version polyfill

TEST=stdcompat-test(package), stdcompat-unittest(host)

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/458957
Original-Revision: 630cd62051bda69ea91c874ab1c2df1b50f6eff7
GitOrigin-RevId: 8820a404dafeff6d507d8931c9b40fad2ea10b87
Change-Id: I63fb58b30f36b9484f9871499e56d3c267480fd1