[roll] Roll infra/recipes [auto_roller] Move update_manifest_project into crashpad_roller
crashpad_roller is the only consumer of this function now that
dart_flutter_roller has been deleted. Keeping it in the auto_roller
means that it has to be covered by auto_roller unit tests, which adds
extra overhead when writing new auto_roller unit tests.

Also, the auto_roller recipe module generally strives to be completely
unaware of the contents of the diff that it commits and rolled.
`update_manifest_project()` was an exception to this, because it's
specific to rollers that update pinned versions of projects in Jiri
manifests. It's not the end of the world to have more specific
functionality in the module, but especially now that it's only used by a
single recipe it's a good time to move it out of the module.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/460683
Original-Revision: d190dbe6e21101c0d2cc0a6fcc00c1eef75e08eb
GitOrigin-RevId: da9174bd4860402993556cb0b46233ff012225dc
Change-Id: I4878e8fb39044390a7199ea2b73b9338aff92c65