[roll] Roll fuchsia [settings] Fix volume earcons via voice
Because of the async code and the recent change from timestamps to
flags, the multiple calls per-stream via voice would flip the media flag
back to the same value, and not register a change. This changes the
flags to a counter.

It also renames the corresponding flag references to "counter"

Original-Fixed: b/175038112
Test: fx test -o setui_service_tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/460601
Original-Revision: 5806553ef86099bcf65ea24dea011141c868e697
GitOrigin-RevId: ce3d61a7c880a0501a62abe81c2b4bae1ab512d1
Change-Id: Iaa18dcb9c02f9d1937f92de0e7be00c515f1ed92