[roll] Roll fuchsia [cobalt][time] Use clock signals to check time is accurate
fuchsia.time.Utc is now deprecated, this change updates cobalt to
instead use the ZX_CLOCK_STARTED signal to detect when the clock has
been synchronized.
Tests are updated as follows:
FuchsiaSystemClock is updated with a handle to a clock. During unit
tests FuchsiaSystemClock is supplied with a fake UTC clock which can be
manipulated during the test.
The timekeeper fake is removed from cobalt_testapp. Instead, cobalt is
supplied a flag that causes it to swap out the real UTC clock it is
provided with a fake one which is always started. This allows cobalt to
complete the clock started check in absence of network.

Testing: fx test cobalt_tests,
fx test cobalt_testapp_for_prober_do_not_run_manually.
Verified that fx test cobalt_tests succeeds on both devices with and
without synchronized time, and
cobalt_testapp_for_prober_do_not_run_manually succeeds on a device with
synchronized time.
Manually verified the fake clock flag is not used by the real cobalt
and that the time reported in cobalt is accurate.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/458054
Original-Revision: 5c3fbf1330acf204fe299fbc7b34c27c3f82582a
GitOrigin-RevId: 7ee41b5139b38cb149d6ab7c025077d06374cad6
Change-Id: I5c9820b3ddf043982558db1177ce53e154f21b9b