[roll] Roll fuchsia [scripts] Fix `fx status` after change in `gn format` behavior
https://gn-review.googlesource.com/c/gn/+/10440 fixed a bug where `gn
format` would output to stderr instead of stdout. `fx status` had code
to work around this, but now that the GN bug has been fixed stderr is
no longer the correct stream to read.

One downside to the new behavior of `gn format` is that, when passed
input via stdin, the dumped json and the formatted GN are now both on
stdout, so we need to trim the latter. However, this option is currently
only used by tests.

This CL updates `fx status` to remove the stderr workaround and add a
different, test-only workaround for stdout.  Additionally, the tests
were updated to account for the more detailed parse tree information now
output by `gn format`.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/460717
Original-Revision: 069017fbf5b76c646525ba106200e0ae722c0491
GitOrigin-RevId: 0c9cd6fa3cee7c14505b13a4f23545628b14cc7b
Change-Id: I9bdefb3a6ba761e330833e74a9a4c69cbfb15ea2