[roll] Roll infra/recipes [build] Use Perfetto for GN & Ninja tracing data
We need to remove the dependency on //third_party/catapult, which
includes the trace2html tool.

This change migrates the tracing logic to use Perfetto, which is a
service that can read a trace from GCS and convert it to a readable
format. As a bonus, we do less work in the recipe.

Original-Bug: 65573
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/458575
Original-Revision: 316c4e80af9bc19edca0545786d61627c4ef2382
GitOrigin-RevId: bca2d0a26f38b5007f7a88d9da894385fcb5ee4e
Change-Id: Id8c2a32ee45c5c00d79f8e6a85b43c43175f2eb4