[roll] Roll fuchsia [netstack] Rewrite DHCP to avoid UDP sockets
Use a packet socket to receive all traffic. This paves the way for
turning off the broadcast flag on outgoing messages, since we are now
able to receive IP packets for which we don't yet hold the address.

Write outbound traffic directly through the stack, rather than using a
UDP socket. This avoids the need to assign the unspecified address,
which causes calls to fuchsia.net.routes/State.Resolve with the
unspecified address to resolve to the local interface on which DHCP is
running. Add a test for this case to netstack_dhcp_integration_test.

Original-Bug: 66094
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/450114
Original-Revision: d0b53b2ce64eb11478db9d92d9e6458f5d0b5225
GitOrigin-RevId: 8706b4748bc6ad884774aa189f5161a95ce9f667
Change-Id: Ic2d89e438beff8ffe4485a3d50a89028de2bcbab