[roll] Roll fuchsia [kernel] kcounters: Convert ifdef'ed atomic to atomic_ref
kcounters provides a way for the kernel to export statistics to
userspace, via a well-known VMO with a well-known format. The
implementation internally is a series of uint64_t counters and
descriptors, which can be aggregated to get a count of events.

The counters were updated atomically under certain conditions (relaxed
atomic add on arm64) but not in others (plain += on x86); convert them
to a lossy-add sequence -- an atomic load, add, atomic store (relaxed);
this eliminates a C++ data race and is iso-performance to the existing
code. The lossy-sequence uses atomic_ref, to not affect the
representation of the field in the exported VMO.

This change also allows removing two more legacy atomics from
kernel/atomics.h, towards fixing bug 47117.

(This language-level data race was detected by KCSAN).

Original-Bug: 47117 Use only ktl::atomic in the kernel
Original-Bug: 61451 Kernelmode Concurrency / Thread Sanitizer for Zircon

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/459854
Original-Revision: 98be88130e1f1a7e82ee224e7aa5b565bfccf35d
GitOrigin-RevId: d80af4bf2f498a4c6d161b11f24cfc39c4e32698
Change-Id: I446f307b801321ac72d68907ed874db0f84af218