[roll] Roll fuchsia [kernel][lib/arch] Have FakeCpuidIo return zeroed data for unknown leaves
It turns out to be too strict a requirement to expect users to
front-load data for every accessed leaf (e.g., with the xsave leaves,
which need to access subleaf 63). We update the behavior to be in line
with arch::BootCpuidIo (as well as the hardware itself in many cases).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/452917
Original-Revision: a6b49c594eb5c4464559885008276c5b26c2a44f
GitOrigin-RevId: bb10be3e1258034316820d2b509518233857b35c
Change-Id: I7a42b6b2b0fbeca1917c231667fbd0b570b051af