[roll] Roll infra/recipes [gerrit] Pin gerrit CLI CIPD version
We use this tool pretty heavily in all of our rollers and fuchsia.py
builders, so it would be bad if we ended up using a broken version of
it.

It's also a very simple tool so it doesn't matter very much whether
we're on the latest version.

Original-Bug: 66363
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/461274
Original-Revision: 0c4f4480f9fa56883af4ef54fe06956fc9c19dee
GitOrigin-RevId: 5ad81863259673b709ddfbcf3831073f860bf7d3
Change-Id: I252f3fdd6ceefdad77434f4dc7c44e0002d32d47