[roll] Roll fuchsia [fidl][llcpp] Fix crash when a server implementation inherits from the interface at a secondary location
When changing the `fidl::BindServer` signature to work with typed
channels, I discovered that the wording around the type of the pointer
to the server implementation isn't very clear. The BindServer functions
were templated with an `Interface` type parameter, which in practice is
in fact the concrete type of the user's server implementation, instead
of the pure virtual `MyProtocol::Interface`.

This was significant in C++ as it turns out the pointer values could be
different when a pointer is converted between these two types, depending
on the virtual table layout. This was the cause of fxbug.dev/42281,
which now impacts both fidl-async and the standard message dispatcher
(fidl::BindServer).

The code previously used tricky casts between the concrete server type,
the interface type, and `void*`. I took the chance to fix this by
introducing a common interface for dispatching a generic FIDL incoming
message.

TEST: Added test case for a server class that uses multi-inheritance

Original-Fixed: 42281
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/459814
Original-Revision: 69fc1871aa8d4b9daaaf7df71e48659c0cc5fbfe
GitOrigin-RevId: 99d2c9b49cc1880b046c613eec779e7b66b3b78c
Change-Id: I3c4683c37adf2527ae2bf04ec5fe15905414ad28