[roll] Roll fuchsia [FDIO] Add Clang static analyzer annotations.
Add handle usage annotations for Clang static analyzer.
These headers should normally be put into a header-only library,
however, due to the fact that FDIO libraries might be used by
out-of-tree users that do not use Fuchsia build system, we have to
put it in the directory that would be included by the existing
out-of-tree users.

In the future, we plan to migrate out-of-tree users to use the
fuchsia build system and move this file into a proper public
library.

Tests: Local build.
Original-Bug: 63743
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/460597
Original-Revision: 7fa6aa56e86ed11457788a2248bc89aa2c6359f3
GitOrigin-RevId: 016bd600c83994391fc9ea840f7848ac943392f3
Change-Id: Ia6c87098e94f3358bdcb37415cb25a43dce30cee