[roll] Roll fuchsia [blobfs] Label VMOs with merkle roots
Previously, blob VMOs were named based on the inode of the blob.

This prevents collisions, which is nice, but it makes it difficult to
figure out which VMO maps to what file. Morever, it makes it *really*
hard, nigh impossible, to figure out which VMO maps to what file in
situations where the device is not accessible, such as when looking at a
crash report. (Inodes are not stable across all devices.)

Change this to use the first 8 characters of the merkle root hash, which
is a stable identifier for the given contents. This introduces the risk
of collision, but collisions do not have any effect on blobfs' correct
functioning, and downstream memory analysis tools can break ties using
other information such as koid.

This results in vmos like `blob-d25911f9` and `mrkl-d25911f9`.

Tested: Trivial change; tested locally.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/457889
Original-Revision: 7a548a1a73b272b21a694f2717bab5d8f5d3ef15
GitOrigin-RevId: 6577762b748b61663aaa313a30926725f3027d97
Change-Id: I61fac6e44dc901710e219f0b7b8c5124fdbaa82a