[roll] Roll fuchsia [fdio] Fix slack in timerfd
Sometimes, waits on timerfd triggered before the timerfd was actually
readable. This was because the slack was not properly set on the
underlying timer: they were intended to use ZX_TIMER_SLACK_LATE, but
this flag was passed as the "slack" parameter (normally, a duration)
of zx_timer_set. Instead, this option should be passed to
zx_timer_create.

Original-Bug: b/175242584
Test: stopped the following from eventually failing:
  fx test fdio-unittest-package -- \
  --gtest_filter=TimerFDTest.RepeatingNonblocking --gtest_repeat=-1 \
  --gtest_break_on_failure
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/460816
Original-Revision: cbfe16ce3af7c09bf9650747cebfa06c4f2bab51
GitOrigin-RevId: 6f44d9530a3d7c2def1e6d8c17b1e7fd4d073fe5
Change-Id: I38ae245a74787bbe52c87aa2b06217f151a702d3