[roll] Roll fuchsia [lapis] Add FTL configurations for Sampler.
This is a reland of a configuration that
was landed and rolled back due to performance regressions.

The performance regressions were caused by a bug in a client
library built by the tool, but initially the regression was
attributed to a sampling frequency set too high. The original rollback (fxr/457656)
was relanded with a significantly reduced sampling frequency, but the
performance regression was still observed in CI (missed on reland due
to only running perf tests, not longevity tests). The persistent regression
despite reduced sampling frequency led to more in-depth investigation, leading
to the underlying client library issue being resolved.

This is a reland of the original CL, with the 60s sampling rate. Perf and longevity
tests show no regression in CPU utilization.

BUG: 66080

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/460835
Original-Revision: def1528d15b769c3ee7b5e5f9c16792cabf3b55d
GitOrigin-RevId: aa53cac84c2e32920a4030c7c3add7cfd74663df
Change-Id: I37ec40b30c750e8f911b1b8eb1e56daf9991769e