[roll] Roll infra/recipes [docs_roller] Delete `do_zircon` property
Based on the comment, it's a vestige from when this recipe ran on both
the fuchsia and topaz repos. But topaz no longer has its own build (it's
still a separate repo but is build as part of the fuchsia build) and it
doesn't have its own docs-roller, so `do_zircon` is always true. So
replace it with a `zircon_target_arch` parameter that specifies the
target arch for which we should generate zircon args, to avoid
hardcoding "x64" in the recipes.

This lets us delete on of the test cases that covered the non-zircon
case. I also deleted some mock test data that wasn't actually necessary
for the tests to pass, since it's passed directly into another command
and isn't parsed by the recipe. (The `gn args` mock data didn't have any
effect anyway because it used `api.json.output()` instead of
`api.json.output_stream()` to mock the stdout of the step.)

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/461715
Original-Revision: 8369cc8914da578672217227a0e288434d5f6fd5
GitOrigin-RevId: 57d0009d82969f7c05ce0debe466073797a37eed
Change-Id: I22621d3a0b71d402fa36e173aa9f80510634d6b1