[roll] Roll fuchsia [flatland] Fix a potential race in Flatland cleanup.
When a Flatland instance encounters an unrecoverable error, it closes
the channel. Right now, the only such error is when a client calls
Present() without any "present tokens".

The actual cleanup is tricky because FlatlandManager holds an
async::Wait associated with the server-side channel, so the Flatland
object cannot close that channel until the async::Wait has been
cancelled.

Prior to this change, FlatlandManager would cancel the async::Wait
from the worker thread, which is not thread-safe, since the
async::Wait documentation explicitly states that it should only be
accessed on the looper it is waiting on. Additionally, the test for
presenting without any tokens could flake if it did not wait for all
of the cleanup processing to occur.

The best way to remedy this is to move the async::Wait into the
Flatland object itself. It does mean that the Wait will have to run
on the worker thread, but this is only somewhat inconvenient in tests
and won't affect real client behaviors. It also allows the Flatland
object to cancel the wait when the client makes an unrecoverable
API call like presenting without a token, which means it can also
immediately close the channel.

Also modifies the FlatlandManager to use an async::Executor instead
of a raw dispatcher to ensure tasks can't access a destroyed
FlatlandManager.

Original-Bug: 64836
TEST=fx test -o flatland_unittests
MULTIPLY: flatland_unittests (fuchsia): 30

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/458274
Original-Revision: b4d540c473e236000a9f10309cc4f2cf8c54b8a9
GitOrigin-RevId: 7078aff423e0e83457b9affee0c45e96731ad287
Change-Id: I04e01d1cfe424bd99ba980ab77280799b83f4128