[roll] Roll fuchsia [inspect][rust] Refactor state towards implementing transactions
Moves the Arc<Mutex> into the State and introduces the LockedStateGaurd
which will be used by the transactions to perform a single lock() of the
State and of the VMO per transaction.

Tested: this reorganizes the code, so existing tests suffice. A new test
for the LockedStateGuard was added in place of the test for the now gone
with_header_lock macro.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/459503
Original-Revision: 56914280799fafa0c84b4c0a14c03109821e0889
GitOrigin-RevId: ae39b38f639e360a5aff0c3d5a131cbf75f3a356
Change-Id: I354431ddaa8414760538c3379228f3d16cfdb42c