[roll] Roll fuchsia [codec] Log stream buffer AllocationError(s)
With this change, I don't know of any video decoder creation
allocation failure paths that lack a logged cobalt event.

If the main input or output buffer collection fails to
allocate, that'll be InputBufferAllocationFailure or
OutputBufferAllocationFailure.  If an internal buffer fails
to allocate, that'll be AllocationError.  If it's some other
non-allocation-related error, that will be a different event.

BUG: b/173512120
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/461416
Original-Revision: e7f7c8c3d247bfbd40b976916e201e89719bcd52
GitOrigin-RevId: 5898950c3bc6d5567c5a4456302f43645ca63d17
Change-Id: I5fa2d0a4facbfd10edc620c328921450b316d485