[roll] Roll fuchsia [kernel][arm] Treat unknown memory range types as reserved.
Currently the ARM startup code will panic if it encounters a memory
range in the ZBI with an unknown type.

We change this behaviour to instead treat the range as reserved, to
give us more flexibility in future for changes to the format.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/462054
Original-Revision: 7a83ed2641ab24ddfbebddaca1e34b36703d90c0
GitOrigin-RevId: c56c9d95624e8d3d92b76d64b185baaed64d3e52
Change-Id: I124c8ac5858c60bcf554f3593e7b64b3df271458