[roll] Roll fuchsia [stdcompat]: Branch fit::optional.
Branch optional and test. This is a copy, before we proceed into
updating so is an accurate polyfill of optional.

The file is not exposed as part of stdcompat yet, and is not built. This
will be updated in a following patch

Original-Bug: 57349
TEST=None

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/461525
Original-Revision: 45701e0fb56163975a44e9ad3add1cc27b120fe0
GitOrigin-RevId: 5d72471ad28b98575bd7d7a45346b3f3d37e5788
Change-Id: I773b567057028a426cff0cd8ce4af605180ef48d