[roll] Roll fuchsia [stdcompat]: Branch fit::variant.
Branch variant and test. This is a copy, before we proceed into updating
so is an accurate polyfill of variant.

The file is not exposed as part of stdcompat yet, and is not built. This
will be updated in a following patch.

Original-Bug: 57349
TEST=NONE

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/461896
Original-Revision: a9072686cec68cc40d52f2240c2f83cd403fee4a
GitOrigin-RevId: adff200ed11db08fa8c76237fdbf0da749f18053
Change-Id: I7874b348936bdb2cf1e737be273f5fe286624fdd