[roll] Roll fuchsia [ffx] Add interface names to link local addrs.
If an address is link local, adds the interface name. If an interface is
unplugged and cannot be shown, for now just shows the index instead of
the interface name.

Hack to get tests working on bots: interface names are given very high indexes
in order to make it highly unlikely that they will be assigned names.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/460846
Original-Revision: a1ec48ee37ac05ceebd6d9b193e0fafe4903e39f
GitOrigin-RevId: 51717e29b8eebe9cf502d18d49eb0a3bf28edacb
Change-Id: Ie8d38a61832ed049357800bef147b9f39c78d7e6