[roll] Roll fuchsia [audio] Fix instance of passing wrong handle type across FIDL
The FakeProfileProvider was using a zx::event handle to fill in for
zx::profile. The upcoming change to validate handle types sent across
FIDL will break this, so remove the requirement that the fake profile
provider send valid handles back. Add a zx_status_t field to the
functions that get profiles so we can still check for errors in the
test.

Test: fx test audio_core_unittests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/462555
Original-Revision: a628aae3ca73160fbc052c3c9b1327663636f9c4
GitOrigin-RevId: cd89d14cf465fb8f59e8144f9e92d59f2050da8e
Change-Id: I7ce22cb1fdd84a1a32bac4dbbb2d4f69a9b37c38