[roll] Roll fuchsia [fx] Add ninja-executable to gn gen args
As of version 1.8, ninja caches mtimes in the .ninja_log file and will
not restat a file if it has an entry in the .ninja_log. This becomes
problematic for generator files such as build.ninja. In a scenario where
the .ninja_log has an entry for build.ninja (ie, after ninja has had to
trigger a re-gen) and a manual run of of `gn gen` has updated the
build.ninja file, a subsequent invocation of ninja will trigger yet
another re-gen of build.ninja.

This can be reproduced like this:

  # Initial build
  fx set core.x64 && fx build
  # Trigger ninja re-gen so there's a build.ninja log entry
  touch BUILD.gn && fx build
  # Manual re-gen
  fx set core.x64
  # The following command will re-gen again before building
  fx build

With a recent gn change[1], gn will now use `ninja -t restat
build.ninja` to force ninja to update the .ninja_log with the accurate
mtime after generation. Gn will do this by default, so long as the user
provides a --ninja-executable to use and the ninja_required_version is
>= 1.10.0 (which is when the restat tool was added). Eliminating the
redundant generation saves 10-11 seconds during the fx build.

[1] https://gn-review.googlesource.com/c/gn/+/10400

Original-Bug: 50061

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/461355
Original-Revision: d399322deea34facaf6cb943d44a0294c34ab886
GitOrigin-RevId: 44861948b8d456d742d8feb5c7afae36a3e3d2c9
Change-Id: I77f1733268fdefc588f2f398900ae4f608575492