[roll] Roll fuchsia [docs][fidl] State that Rust bindings do not support struct defaults
We do not plan to support default values for struct members in the FIDL
Rust bindings because there is no easy way to implement it. A method
returning a struct instance (via std::default::Default or otherwise)
would need to pick values for the fields without defaults, such as 0 or
`Handle::invalid()`. This would go against good Rust style, which seeks
to avoid these in-between invalid states.

There are alternatives, such as only supporting defaults if _all_ fields
have a default, or using a macro instead of a method. We are
choosing instead to explicitly state that defaults are not supported.

Test: docs change
Original-Fixed: 8013
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/462676
Original-Revision: 3eb27f57097998379ea5248ffb4f66299351a2bf
GitOrigin-RevId: 128febf9e2e9c7d7ba7a23e64aed1ba596aaec1c
Change-Id: Ieaac47a8d2aa17e84cd2158470cdac5f89878375