[roll] Roll fuchsia [build][alln] Add vdso_toolchain.gni and update libc_toolchain.gni
Add a new .gni file that is used to compute the exact toolchain,
target and output files for the Zircon VDSO, based on the current
build variant selection.

This also updates the definitions in libc_toolchain.gni to make
them more consistent with the ones in vdso_toolchain.gni, as
well as //build/config/BUILDCONFIG.gn.

This prepares moving the sysroot to the Fuchsia build entirely.

+ Ensure that the VDSO is never built with GCC< since this is
  already what happens in the Zircon build. It turns out that
  the VDSO sources can no longer be built with this compiler,
  which complains about missing function attribute specifiers.

+ Ensure that the SDK sysroot's base (non-instrumented) libc.so
  is built with the build variant as the system C library
  (e.g. with "gcc" or "thinlto"). This is necessary to avoid
  conflicts later when generating Fuchsia packages that
  contain binaries that depend on //sdk:core indirectly.

Original-Bug: 60613
Testing: Should not change the build, except for the SDK sysroot's
libc.so when using a "gcc" or "thinlto" variant.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/459202
Original-Revision: cafd70b2d2df7b77671909ce9467b92fcf149cfe
GitOrigin-RevId: b5896e8cd62805a6fc9b899dd0989b71416c76f0
Change-Id: I30c21bbaabc9209d205e40b87b177ea2e848876c