[roll] Roll fuchsia [chromium][vulkan] Convert to Vulkan cmx shard.
Add a new cmx shard that's similar to test-application.shard.cmx, but
doesn't add the vulkan feature to the sandbox. This is useful for test
components that don't use Vulkan directly, but instead launch other
components that need to use Vulkan, and so need to pass the Vulkan
services down.

This cmx shard is useful to ensure that all tests in this situation have
access to the set of services that the Vulkan driver needs to operate.

Original-Bug: 64508
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/461761
Original-Revision: 4f611a6d57f2fa9c1d32d8b754a096fb11a0c788
GitOrigin-RevId: 2ab170c5f1d62858a533ac0ccd3ba4cf364e1818
Change-Id: I55e8320051cf8fb9d1bb7aa5b33bb6d512506017