[roll] Roll fuchsia [fshost] Fix logging to serial
Opens handle to log write service and reconfigures syslog to use that
handle for logging. This is a short the fix for a bug where in on a
board with userdebug build, no logs show up on serial.

Test: Monitored serial for the messages.
Original-Bug: 66263

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/461976
Original-Revision: aa3dd4441ab252eb8ad021d9fd1b170c307a997b
GitOrigin-RevId: b73ef1e8e3448f0787f6b43885e21e790e349136
Change-Id: Ia581e175ef9e3f678f6b25339aaba4d76f508512