[roll] Roll fuchsia [harvester] Gather static annotations for control panel
We want information about the device for a new FDT feature. We are
already using the build-info data for the snapshot fuchsia version, this
just reads some of the other data.

This is to unblock us until the feedback service exposes a table with
the values that we want.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/450236
Original-Revision: bc7a51ad69a806454926acf3938d529f9b66d0f7
GitOrigin-RevId: 63db5b62e6875a05f570c985555c8fdead1b8ba5
Change-Id: I275a2dc1135cc11ceea01fca50d1e7c234904d92