[roll] Roll fuchsia [kernel][vm] Fix bug in commit offset for batch page requests
A commit on a pager-backed VMO (or its clones) typically generates a
batch page request for the entire range being committed. However if the
range being committed already has populated pages, the request gets
broken up into smaller batch requests, one for each contiguous range of
absent pages.

There is logic in VmObjectPaged::CommitRangeInternal to reprocess a
range that was freshly populated by the pager, i.e. call
VmCowPages::CommitRangeLocked on the range twice. The second call allows
VmCowPages::GetPageLocked to perform any COW copies of the newly
populated pages if required.

For batch requests on ranges that already had some populated pages, we
were incorrectly advancing the counter tracking the committed length as
soon as we encountered a populated stretch. This caused us to not
reprocess the unpopulated range after the page request had been
fulfilled. So we would end up not making copies of such pages for
commits on the clone of a pager-backed VMO.

Note that this bug did not lead to any incorrect behavior with our
current pager implementation, because we would still correctly copy out
the page if the clone wanted to write to it. We were just not doing it
early at the time of the commit. The only observed effect of the bug was
that a clone might see fewer committed bytes than expected after the
commit.

Test: core-pager
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/462794
Original-Revision: 7f856507d410b8b432f702fe0a637472201fefc9
GitOrigin-RevId: 79e242cc19c2cc4bdc3bf83e8c8276843d194147
Change-Id: I1c22e4d9515c1d53e5bab75cfe0d9a397bbb7749