[roll] Roll fuchsia [build][alln] Remove legacy_sysroot-$cpu.json and gen/legacy-aux-$cpu.json
This CL removes two files generated by the Zircon build and parsed by
the Fuchsia build, since the latter no longer needs these.

Note that this CL drops the number of targets in the Zircon build to
849 (from over 2200), mainly because the latter used to build all
variants of the C library, just in case one of them would be needed by
the Fuchsia build. After this CL, only the C library variants that are
really needed (based on the dependency graph) are actually built
instead, so there is no loss of feature here.

+ Remove //zircon/system/ulib/BUILD.zircon.gn since it no longer
  provides useful metadata. Note that the boot-options line was a
  mistake and should not have been added here (despite comments
  explictly warning about this!). It was also added to the BUILD.gn
  file in the same directory, and that addition was correct.

+ Remove the zircon_aux_manifests and zircon_boot_manifests variables
  from //build/images/manifest.gni since these lists are now completely
  empty anyway.

+ Remove //build/unification/images/* which are no longer needed.

Original-Fixed: 60613
Testing: Should not change the build.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/459774
Original-Revision: 01fa0fb2619f837676c238fde522e45680745bf1
GitOrigin-RevId: 51b02689ec9579c2d8ce7915cb40ec7f0def4ee0
Change-Id: I64371c95dc721957b065d01326493a5bd96514fa