[roll] Roll infra/recipes [affected_tests] Add fix that was removed in revert fxr/461447.
The revert reverted a CL that changed the script to look for phony rules
instead of stamps in the ninja output. However, it also included some
fixes to the detection of affected tests, which was missing previously.
This CL adds the fix back so we should be able to detect more affected
tests.

Original-Bug: 61759
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/462603
Original-Revision: 1205eac2a9a50825dbc76d7bd13d6fa90851a326
GitOrigin-RevId: 9aac825526e07232e5b41fe7f9ef6ae5266238fe
Change-Id: I86e55c0e31cf1960931aacbf4e28d6f776a60019