[roll] Roll infra/recipes [buildbucket_util] Fix test on Mac
This test was previously failing on Mac due to some weird unicode
errors. Part of the issue is that emojis are encoded as two "surrogate"
unicode characters on Mac rather than just one unicode character.
However, even when I tried switching to less weird unicode characters
like "å" and "•", the summary still contained one too few of the
expected characters on Mac.

I don't really think it's worth fixing properly, since the main purpose
of the unicode handling is just to ensure that the resulting truncated
summary is valid unicode that won't break Buildbucket when we try to set
the summary_markdown. We don't need this assertion for that.

Original-Fixed: 66536
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/462716
Original-Revision: 740fdc5c19d0bf331f46d299b1501152545ec80e
GitOrigin-RevId: 35e5f70c0b76bf0aad8acf2802236f6e5c06958f
Change-Id: I6fee90113d54e22c510698bfeb45774322bd037d