[roll] Roll fuchsia [fdt] Ensure temp directory exists for unzip
The way unzip -d works is that it will create up to one level of
directory that does not exist. If you tell it to extract to a/b/c
and either a or b does exist then the operation will fail.

Previously our CIPD download would be something like
fuchsia_devtools/macos/macos.zip

Now it is just macos/macos.zip

To minimize the amount of code difference between the old/new I kept the
target of the unzip to still be fuchsia_devtools/macos-extracted-temp,
but it turns out that when the download toplevel is macos instead of
fuchsia_devtools, we are missing that one level of directory so the
unzip can fail.

This adds commands to ensure the directory exists before extracting to
it. This is benign because all of the code ensures these temp
directories are cleaned up anyway if something goes wrong. This simply
ensures that we don't ask unzip to create too much nesting.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/462294
Original-Revision: fc461a7f7dc08a5bd8af2fa09b6181002cc8b51e
GitOrigin-RevId: 72a2f0ce7664f953b6d3237bdaf243c6e2c44cbb
Change-Id: I5a1597441b3fd3c3dd2a14f5402d57dfa151e8cf