[roll] Roll fuchsia [cts] Verify Deps are in released in an SDK
This CL changes how the CTS verifies that a dependency is released in an
SDK. Previously, the CTS would look for a {dep}_sdk.meta.json file. This
file may not exist at the time of CTS verification, and it only verified
that a dependency was an SDK atom, not that it was actually released in
an SDK. This CL introduces an implicit dependency on public/partner SDKs
to the CTS. These dependencies will generate all the SDK manifests for a
particular toolchain in the out directory. GN will wait for these
manifests to be generated before running the CTS dep verification
script.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/453857
Original-Revision: 2e32e4d0d777f817f6d0c6d0fa9fe9636f18dada
GitOrigin-RevId: f0283eafc9a5894b180dc0ce03e3df426e5baa78
Change-Id: I7609bb2f67eec271dec8986f8334ccae35c5d0f6