[roll] Roll fuchsia [storage] Add support for oldest version to Cobalt metrics.
To make this work, I wanted to make the cobalt collector private, which
meant removing support for custom metrics which isn't currently used. We
can add support back in later if we need to.

I looked at adding a unit test for the Blobfs change but it looks to be
non-trivial. Since the change to Blobfs is trivial and we do not seem to
be testing any of the other Cobalt metrics from Blobfs I figured it
isn't worth the effort.

Original-Bug: 64651
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/460911
Original-Revision: 8dd460c8373ba8ebdf44ea017137d36fe3f51800
GitOrigin-RevId: ac167fc7da7c1023e8465b1396bb8709c7a985a3
Change-Id: Iee38637df3588cf77010ad509cb4eabaa487752c