[roll] Roll fuchsia [botanist] Split up EOF tefmocheck
Instead of having a single generic "failed due to EOF" tefmocheck for
botanist, check for the two separate EOF-caused failures that we've
noticed: serial read failures and package repository setup failures. We
suspect the former is fixed and the later is still happening, but we
can't be entirely sure as long as they're attributed to the same test
failure.

Original-Bug: 65073
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/460295
Original-Revision: a703df151e713809958fbac0d9591064231747f2
GitOrigin-RevId: b8297d913dad269fb9892d57bbfd74283867b547
Change-Id: I6a29e08f7293a61a46fb5496f09961ef03cc1040