[roll] Roll cobalt-registry [managed_os] omnibus metrics change and cleanup
Refactor of metrics with multiple dimensions to support consistency in
reporting when adding new metrics in the future. Specifically, any
dimension after the first now reserves its 0-index code to mean
"Unreported". A prominent comment is added explaining this situation and
reminding contributors and reviewers to consider this when modifying
this file. We are OK with breaking the existing reports for consistency
later. For the changed metrics, the dimension at the 1 offset is not
moved in attempt to keep as much data valid as possible.

Dimensions with inconsistent casing between metrics definitions
triggered a bug in the source generator. Dimensions are renamed to
PascalCase globally for consistency.

This change also adds four new metrics:

* When an update check is performed, we record the disposition of the
check, how the check was initiated, the result of the check, and the
result of the update. This is reported in a per-day linear histogram
of the sum of performed checks.

* When an Omaha event report is interrupted and therefore lost, we
record the type of event that was lost, and the result of the event
being reported. This is reported in a per-day linear histogram of the
sum of lost events.

* A metric to report histograms on the number of devices that have
seen N consecutive failures to apply an update, per-day.

* A metric to report the number of devices that have seen N consecutive
attempts to apply an update, regardless of disposition, per day.

The `update_duration_from_first_seen` metric is augmented to accept a
dimension of the clock type reporting the metric.

This change annotates all the unannotated `window_size`s. It also
extends all `expiration_date`s to 2021/11/01. Finally, I took the
liberty to fix some typos and grammar in comments.

Original-Fixed: 65795
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/cobalt-registry/+/463196
Privacy-Approval: Steve Fung <stevefung@google.com>
Original-Revision: 23277b256c70564573bda23e11aa2318f60f7cea
GitOrigin-RevId: d91f060c45edc1d641e27615a04d707bdeb652e0
Change-Id: Ic68cbd9187fe16c2edcf0855cb8c558b4cc93f73