[roll] Roll fuchsia [power] Prepare node config targets for bootfs
Update each node config-data target to also build resource targets that
will include the node config file in bootfs. This is a no-op initially
but once the node config target is included as a dependency for a ZBI
target (coming in a later CL), the config file will be available to
power-manager in bootfs.

Test: first pave changes then `fx test power-manager-tests`
Test: Verify Astro boots and power-manager is running
Original-Bug: fxbug.dev/65893
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/460454
Original-Revision: 6187e5734dd93c0337574ded7133a74c6d45ca52
GitOrigin-RevId: c19d01957a5fe813f9e965fdc4645246230c9afa
Change-Id: I641ba197885dbc2065e048d4bb8191a28c31f5b6