[roll] Roll fuchsia [build] Fix libc and vdso toolchain selection.
The code in {libc,vdso}_toolchain.gni was missing a check against
the selector.label field, which resulted in incorrect variant
selection in certain cases.

For example, `fx set ... --fuzz-with asan` would end up creating
the following definition in args.gn:

  select_variant = [
    {
      variant = "asan-fuzzer"
      target_type = "fuzzed_executable"
    },
    {
      variant = "asan"
      label = "//sdk/lib/fdio",
    },
  ]

The second scope is required because the fuzzer runtime requires
a non-fuzzing fdio library to work properly. This definition
unfortunately forced the libc_toolchain.vdso code to select
the "asan" variant for the C library, which was installed
as lib/asan/ld.so.1 in packages, instead of lib/ld.so.1.

This resulted in failures to launch /bin/bootsvc when launching
system images, e.g. in the emulator, because the binary was
compiled to look for the latter path, which didn't exist anymore.

This fixes the issue for both the C library and the VDSO, though
the latter was not impacted in the case of fuzzing.

For testing:

  fx set core.x64 --fuzz-with asan
  fx build
  fx emu -N --headless

Before the CL, the system would reboot due to the inability to
load /bin/bootsvc properly. After the CL, it starts normally.

Note that there are still issues with Go-based fuzzers not
having the right lib/asan/ld.so.1 library in their package
manifest, but this will be addressed by a future CL.

Original-Bug: None
Testing: See above.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/463834
Original-Revision: 24280e640f214215b6bb7441b5478dfa99a023e2
GitOrigin-RevId: fc32594f39ec88e4f88bff810799851c2a440a1d
Change-Id: Id877d4e8ee29a298b69c31ecbdd8566f64216cff