[roll] Roll fuchsia [build] Fix ubsan variant build.
The "ubsan" is broken. The main issue is that libclang_rt.ubsan_standalone.so also
depends on libc++abi.so, and indirectly on libunwind.so, and this
is not properly reflected by the runtime.json file provided with or
Clang prebuilt toolchain.

This fixes the issue by patching runtime.json after reading it,
similarly to what is done for libclang_rt.asan.so which has a similar
issue.

Original-Bug: 66509
Testing: "fx set core.x64 --goma --variant ubsan" then "fx build" should succeed.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/463835
Original-Revision: 27d4e251ef808128b51140747011036c1349815d
GitOrigin-RevId: 5551cdea59be554c14ad7ba231b167ab9f96ab09
Change-Id: I90d005cde9fb330d74142664945110b447b5752e