[roll] Roll infra/recipes [led] Delete led recipe module
led is now installed in the ambient Swarming environment and kept pinned
to a well-tested version. So roll in the recipe engine change that uses
the ambient led version instead of installing it from CIPD
(crrev.com/c/2591797), and delete the "fuchsia/led" recipe module that
we created specifically to pin led to a known good version.

We'll soon no longer be reliant on led once recipe testing with
recipe_bootstrap is rolled out, so it's not a big deal that we won't
have control over the led version we use.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/463854
Original-Revision: 01104e9cbbefaeb1497403e496655918ad7fb416
GitOrigin-RevId: 284dba0aeb8046842fd0cbce274ac4813bacf56b
Change-Id: I9d97ab9af0b4709d77ccf5059ac2d8b561964a39