[roll] Roll fuchsia [lowpan][lowpan-spinel-driver] Bring interface online if provisioned
Currently, if the LoWPAN device is provisioned and the system reboots,
the interface will remain in the disabled state, requiring someone to
do the equivalent of a `ifconfig lowpan0 up` to bring the interface
online.

This change fixes this situation by changing the default initial state
of the interface to enabled. In the course of implementing this fix,
several other issues were identified and corrected, such as some
confusion between the `enabled/disabled` and `online/offline` states
in `tun.rs` (and some other parts). The way addresses propagate from
the NCP to netstack was also made to be more robust and to also
exclude the internal mesh-local addresses that are not for application
use.

Test: fx test lowpan-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/460855
Original-Revision: f76cca0c9b34360cc2fef9fd53a6c35004f0c5da
GitOrigin-RevId: 6a118aad04d5fc2876ac421917c79ab87713291c
Change-Id: Ic6146d942140a1ba803f28b6eaec4953dbca6911