[roll] Roll fuchsia [settings] Replace SettingResponse with SettingInfo
This CL completes the replacement of SettingResponse, which was part of
the switchboard, with SettingInfo, which resides in the base of the
service. The original CL, fxr/460716, duplicated the enum as part of a
larger effort to refactor the changed notifications so it didn't want
to make huge changes in the system. This CL finishes the move + rename.

Original-Fixed: 66338
Test: fx test setui_service_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/463254
Original-Revision: e1141cb5b94a220c3d5afb5a955d8400ddf81e15
GitOrigin-RevId: e36f7c7576d6b0d888333354735775cf2c8ce20f
Change-Id: Ifb6986ec3d43fff7df87f547f33d025dca61a534