[roll] Roll fuchsia [ftl][ndm]: Reduce bad block to debug.
This is just reporting the list of bad blocks found in the device.
This is expected, as there is a block reservation that replaces blocks
gone bad.
Essentially this is the list of bad blocks reported by the parent
device, or treated as the factory bad block.

Original-Fixed: 66307
TEST=CQ

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/461523
Original-Revision: d0c9a1a3e6cca1c4f737f68ab7952597e0bf4884
GitOrigin-RevId: 1e7d4abb378d8955ad059415109672eee9fbda02
Change-Id: I77befc7766e974a35e2004b2e41f01f7ec48b985