[roll] Roll fuchsia [vfs] Improve vfs header file.
Move OpenResult definition out of the VFS class to make the Vfs class
easier to follow. The OpenResult structure is the same.

Rename vdircookie_t to VdirCookie according to the style guide, and give
the members default values. Use the constructor rather than the custom
Reset method, which is now deleted.

Fix some undefined behavior in the microbenchmarks test. In two places
it forgot to initialize the vdir_cookie and depended on the value of the
dir cookie not changing between loop iterations even though the variable
was declared inside the loop. The variable is now properly initialized
the first time outside of the loop.

There should be no behavior change other than the test fix.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/463962
Original-Revision: d5acc9e3c6b8826ed9d165b7409b2965f885bac8
GitOrigin-RevId: b1d7e88c7e3e90623d664d117b007f258aa0b2d7
Change-Id: I0c76c1f3622b2cba72fb5e9a3476ec9d48636c9b