[roll] Roll fuchsia [omaha-client] Add metrics tracking lost events
Update events are sent back to Omaha by the client, but this is on a
"best-effort" basis. Any issues resulting in the event being interrupted
or ignored do not result in retry. This change sends these updates to
our metrics reporter, which will ensure delivery of this information.

Added a new test to specifically track the production of this kind of
event, and updated an existing test that required it.

This must not be submitted until fxr/457834 is complete.

Test: fx test omaha-client-lib-tests
Test: fx test omaha-client-service-bin-tests
Original-Fixed: 60594
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/458218
Original-Revision: c83f665b2d04be42cf85ae8b157b1e5c72886739
GitOrigin-RevId: 7a3c82b7359a33312bb8a8f744d7e85169bdbc9a
Change-Id: Ia3488db20e3a15a0d72133c29ff7818d8a61e6e9