[roll] Roll fuchsia [magma][msd-vsi-vip] Increase hang check timeout.
To be able to complete the OpenVX-cts on some of the slower
devices the hang check timeout need to be increased.

The current timeout is 5000ms. On the slower devices we need a
6000ms for the tests to be able to run successfully.

Tested using OpenVX-cts on device.

[ ALL DONE ] 7342 test(s) from 81 test case(s) ran
[ PASSED   ] 7342 test(s)
[ FAILED   ] 0 test(s)
[ DISABLED ] 6985 test(s)

Full details of the testing are in the bug.

Original-Bug: 66894
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/464974
Original-Revision: 4b5772fe1c028f38b067875dc31460cf5f408bd8
GitOrigin-RevId: 1de93f1a6e81189be6329854ea0d40f581f55e39
Change-Id: I55b388a5c06e41ddd12f0d3310f65e2c113abe0a