[roll] Roll fuchsia [check-licenses] Reorganizing textextensions list.
strictTextExtensionList in the config file is a list of filetypes that
should contain copyright information at the top of the file.

textExtensionList is a list of filetypes that can't explicitly contain
copyright info at the top, but they should still be covered by a valid
project-level license.

Logically, the strict list should be a subset of the other list, but
previously they were handled as entirely separate lists. This removes
the duplicate entries from the textExtensionList, and modifies the code
to do the right thing.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/465574
Original-Revision: 8a4b7aad46a7396129a07dd5453919178318d094
GitOrigin-RevId: b2445d66221458b4c892649aed07e4004c1c7183
Change-Id: Ic7d9a34c0e145ddb23521bbd30c3679605ff3c10