[roll] Roll fuchsia [ffx] update IPv6 scopes as they change
It is important to try to preserve scope information, which is tricky
because it is host-local. ffx will discover un-scoped addresses from
sources such as netstack queries over Overnet. In those cases, it is
best to preserve any previously discovered scope for the given
address, which likely came from a host-side discovery protocol such
as mdns. If the incoming merge has scope information, that
information should always be preferred, as it may be more fresh, such
as in the situation where a device restart triggers the interface id
to change.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/465478
Original-Revision: 092d53a6a919766f13aafc6dc8886b2b1a4299c1
GitOrigin-RevId: cbcb3b3787139aafd2906f3874cee94f1f2b4765
Change-Id: Ibbdcb54685bf1d21d7230314ea3610cfb9dd4d6a