[roll] Roll fuchsia [fx host-tool] ensure build outputs to terminal
As host-tool is often used in a captured output capacity, the output
from a build would be disruptive. In order to avoid this problem, the
tool now checks to see if there is a controlling terminal, and if so,
redirect output to it. In the case where there is no controlling
terminal, or no terminal on stderr, the automatic build is skipped,
and an error is printed as if --no-build was passed.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/465402
Original-Revision: e9dc78e3093ced95da2bf31c515fb562579a45d0
GitOrigin-RevId: 364c1f2618b5b2e7a85d87c0be45ab9bda1db677
Change-Id: I0c96281cc5816bde849331a40d2381fd98063a1f