[roll] Roll fuchsia [build] Remove duplicate zbi entries from images.json
A first CL [1] mistakenly introduced duplicate zbi entries in
images.json. A second CL [2] tried to correct the issue, but
introduced another issue and was thus reverted.

This CL is an attempt to fix the issue for good. The root of the problem
being that the implementation of the zbi() template in the Zircon and
Fuchsia build are slightly different, i.e.: the Fuchsia version expects
its caller to set the zbi metadata properly, which also allows adding
multiple entries to images.json from a single zbi() target call. The
Zircon version does not support this, and instead would create a single
entry directly within the zbi() template implementation.

- No more duplicate zbi entries in images.json, by removing the
  metadata merging from the zbi() template, and adding it to the
  zbi_test() template instead, to ensure kernel zbi tests are
  properly listed in images.json.

+ This fixes the "cpu" value for some zbi test entries, which could be
  empty in certain cases.

+ This removes an erroneous ".zbi" prefix from zbi test entries' "name"
  value (e.g. 'zbi-hello-world' instead of 'zbi-hello-world.zbi').

+ This fixes the name and path of qemu-phys-xxx zbi test entries
  (e.g. 'qemu-phys-hello-world' instead of
  '_qemu_phys_test.qemu-phys-memory-test.zbi_test.zbi')

Original-Bug: None

[1] https://fuchsia-review.googlesource.com/c/fuchsia/+/435654/
[2] https://fuchsia-review.googlesource.com/c/fuchsia/+/463313/

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/465974
Original-Revision: d5944e2b7a2bed569049cf3e05046bd0592553c1
GitOrigin-RevId: b8989e56a5d874d7c20dd073412afb76d3fa9341
Change-Id: I60e0d19ac642988ab60ae2c96c59b0dd205a5ace