[roll] Roll fuchsia [build] variant_toolchain_suite(): Add prefix_configs parameter
This allows one toolchain definition to provide a list of configs
that will be inserted before the list of default configs that apply
to all linkable targets in a given toolchain() instance.

This is required to move the kernel to the Fuchsia build, in order
to let it override system C and C++ headers with its own versions.

Original-Bug: 66830
Testing: Does not change the build.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/465978
Original-Revision: 03beec1a7a7be2e1d1dec6ca23e81261ea396b3c
GitOrigin-RevId: 571ecf285748bf94b1f1470ba8617cf656a11878
Change-Id: I75f2959d62a48f345172fc03452c0d4f6370e2e8