[roll] Roll fuchsia [check-licenses] Reorganizing testdata directory.
filetree tests all need two files to be present:
 - want.json (representing our expected filetree)
 - config.json (used to generate the real filetree)

Previously I would have to dirty up the root "filetree" test directory
with these files for every test, but that would get difficult to
maintain. This CL moves those files into a subdirectory with the same
name as the test, and creates a new subdirectory called "root" which is
used as the basedir for the filetree being tested.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/465575
Original-Revision: bbcbf16668858e9af48dafeafe5d19416dae0451
GitOrigin-RevId: bf807b1ac82f662cb6d924d5ea68354058aeb22c
Change-Id: Ie8e63f0089324a69aec94c1fc58c1eb228963f5f