[roll] Roll fuchsia [rust][trace] More robust testing for the crate
This crate is a thin wrapper around the C++ tracing API. The wrappers
provided are not consistently tested, and testing prior to this change
was effectively "not crashing is success".

This change adds more robust tests, based upon the C++ test fixture. The
caveat is that the alert function/macro are not (yet) tested. The
`alert` feature is designed to allow signaling an event to the trace
engine itself, which may optionally be mapped to an action. These
signals do not result in generating records, and there isn't a clear,
direct way to add this.

The new tests are implemented using a `rustc_staticlib` target that
tests various aspects of the provided wrapper functions and macros.
These tests are exported with a C ABI, and are then called by a C++ test
that calls into them within the context of the testing fixture.

Old tests are removed as they don't add anything.

Original-Bug: 64487
Test: This change introduces only testing, which passes.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/462135
Original-Revision: 8c7d7e869f8965d46be9096d2a38ed922dce66d3
GitOrigin-RevId: d84e2f573ea75071df04a9f312a6b254db14531f
Change-Id: Ic93f0c7f25810d3c9d5466e31171c540042c1497