[roll] Roll fuchsia [system-update-checker] Prevent installing updates while current slot is pending
In order to guarantee the system is healthy during an OTA, we need to block OTAs on the
current system being committed. While in the update flow, we check if an
update's CommitStatus is Pending via polling CommitStatusProviderProxy.
Timeouts are interpreted as a pending commit.
If so, do not start a new update until the CommitStatus is set to Committed.

Added unit tests and integration tests which verify
behavior.

Original-Fixed: 64603

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/460924
Original-Revision: 9439f8e1c9c9d2439b392c491440a68bbfa4fd1d
GitOrigin-RevId: 21991d517238a69ada9cd6b6985a3924af1059b3
Change-Id: I7ffd6dd3ab1a668ec743dc70f556776830642862