[roll] Roll fuchsia [kernel][x86] Assert that a page table is not free before walking it
We are tracking a potential use-after-free in the x86 MMU code; while
walking page tables, we've found pages in the VM_PAGE_FREE / pmm free
state. It is possible commit 5654d72 fixed it, but to gain confidence
in the fix, add asserts that would catch the same class of bugs
in non-kasan builds.

The asserts check:
* Before reading a page table entry from a page table in the Query
  and Unmap paths, check that the page table page is not free.
* When Querying down a page table, check that next page tables are
  not free.

We use the 'is not free' state rather than 'is mmu' since there are
some early boot page tables that are not marked as MMU pages.

Original-Bug: 66978 kasan panic while running core tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/466579
Original-Revision: 1ace2ed342281c5d2b459a805cb59151d3196b7c
GitOrigin-RevId: c4cc4238483e50a3f13247cc079db973fac6a7eb
Change-Id: I7c5f14a6ad252673758666f6dd3648a0e46fcdbe