[roll] Roll fuchsia [botanist] Teach GCE targets to use existing VMs
Eventually we will want to move instance creation out of the task in
order to save time. This prepares the target for that by having it
attach SSH keys to an existing instance when a name and zone are passed
in by the calling recipe. The recipe changes have not been made, so this
should be a no-op.

Original-Bug: 9127
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/466580
Original-Revision: 59e474896a50dedf17ddd157885da818d8dde8ac
GitOrigin-RevId: 85d47bb03e66c862eeec50825e4eb3aa6e69c01c
Change-Id: I4d98814e7ed90b380b57face19ef78b7ddbd16e6