[roll] Roll fuchsia [system-updater] Add a timeout to cobalt flushes
We've seen some test flakes in which we appear to hang on Cobalt
flushes, though thankfully this won't be an issue in OMCL-based setups
(because they allow a reboot with PerformPendingReboot as soon as we
enter the WaitToReboot state, which happens _before_ we flush Cobalt
metrics).

Add a timeout to that task to ensure system-updater makes progress,
even if we drop metrics.

Original-Bug: 64452
Test: refactored cobalt flush into a separate function, added unit
test.
MULTIPLY: system-updater-bin-test

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/466635
Original-Revision: 91f4bd84db874b5693f4f8040e4f5a39facc701b
GitOrigin-RevId: 3d2a4bba2180dd3fc61c81e4de6a90288668f574
Change-Id: I059d15df39f334f7be76a406a05a5ad7da47a745