[roll] Roll fuchsia [ffx] Clean up ffx target help and add examples
The goal is to have a comprehensive help and examples so that the user
is able to understand the commands and operations.

Also provides some guidance for devs who are contributing plugins to
follow a similar structure and language for the help text.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/467074
Original-Revision: 969ad873b0f2f3134ce6c345bc1bc239fca3fdd1
GitOrigin-RevId: 38e167da630cdedf18b80d18fa1bb7422d52db25
Change-Id: I9ed76f1731c38fa7f785cbe93e693d3bb7a6bd22