[roll] Roll fuchsia [input3] Document the disconnect behavior of the Keyboard FIDL protocol
The Keyboard FIDL protocol may disconnect KeyboardListener as well if
the Keyboard connection is closed.

This is a convenient behavior as it allows us to know when a client has
given up on the connection. But it should be documented as it means two
ostensibly independent protocol connections have intertwined lifecycles.

Original-Bug: 66921
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/465396
Original-Revision: 5636d5cc11c02355cb56601967165795c3a71858
GitOrigin-RevId: 0562d7e18e6c43a483aadb7115bb9284ed0bd2f4
Change-Id: I6543c76fb01ca7dbcafb76ebab96fe38bdf0b3fb