[roll] Roll fuchsia [components] Avoid validating program.binary for runners that are known not to care about it
Introducing new runners such as these in the future will require extending this hack.
We'll need to come up with something more robust than this.
in the meantime, this check has been super helpful and I'd like to keep it to as few false positives as possible.

Also introduce a suppression marker because you always need an escape hatch.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/466934
Original-Revision: 57cf628993303007f68cda59293b6012ccd6587f
GitOrigin-RevId: a69750ddad5f4b563d92f058edcfe72a46f116b3
Change-Id: I0fa5ec2f6583f55c48bc6ddce828855ba894acd2