[roll] Roll fuchsia [component_manager] Reland Enforce Component V2 Capability Allowlist Policy.
This change enforces capability allowlist policy enforcement in Component
Framework V2. This change will prevent capabilities that are restricted
within the RuntimeConfiguration from being routed to realms that are not
defined in that set. Routing will only be blocked if a capability
attempting to be routed falls within that set.

This change fixes the flake in the storage capability testing.

Original-Fixed: 61864
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/463191
Original-Revision: 090fcdb47cf8fa7fd3e521cc58c5cec4a67e45ad
GitOrigin-RevId: 033284bf3b22d437927b1906a1808fa38960935a
Change-Id: I9859cfc64754183bcb2e692f1fd2383c3e74f8e9