[roll] Roll fuchsia [debugger] Enhance "ss" command.
Detects indirect calls in the disassembler and marks them. This
information is used by the substatement call extraction code to list
these in the options for "ss". This allows virtual function calls to
appear in the list.

More greedily consumes line table entries for the code following the
last inline call of a statement. The case I found had a statement
consisting only of a sequence of row entries for the same line,
intersperced with "line 0" entries. This previous algorithm only got the
first line entry, now we find all of them in the range.

This also limits the searching to the code actually in the function to
prevent it from going off the end and trying to read garbage.

Fix a bug with address range finding exposed by the new substatement
test, and add a test for it.

Original-Bug: 67061

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/466775
Original-Revision: ba10816c27168ac9dee5e41411acc16be281cb4b
GitOrigin-RevId: 72f7846c9d2424bbbd7f06f2db9201ef66d3e722
Change-Id: Idb224c5d8934d575f324c7d3bf3b94300d03b603