[roll] Roll fuchsia [debugger] Clear debug_agent's states on disconnecting
This creates a more consistent behavior between the debug_agent and the
zxdb so that further reconnecting will work as expected.

This patch also allows a debug_agent to run as a deamon in the
background as multiple zxdb sessions could connect in sequence.

Test: disconnected & reconnected the same zxdb session multiple times.
      ran multiple zxdb sessions against the same debug_agent.

Original-Fixed: 64872
Original-Fixed: 60034
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/467734
Original-Revision: 03d52d91edfdffa516ed6144f33b2edd90681d19
GitOrigin-RevId: 832784ed4cd660e0e54cd58454df00b31a76f30e
Change-Id: I385bd0f2f4423cd67e35627fc5284868f6558747