[roll] Roll fuchsia [events] Add support for individual sync/async events.
This CL introduces the event mode property in manifest files which
currently consists of async and sync. Those properties can be routed
across events. They're a set (of currently a max of two items). As you
route an event, we take the intersection of that mode set each step of
the way. This signifies whether or not a given component can support
registering to async and/or sync events for a particular event name.

Additionally, this CL introduces the notion of an
EventSubscriptionRequest which lets individual clients specify whether
they want sync or async events for a given set of event names. A single
event stream can receive a combination of sync and async events but only
one of each for a given event name.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/456617
Original-Revision: 4ef4d5a9d3fa452b51ba726b46594e6b61ddf4d3
GitOrigin-RevId: 8f5265a566b152ea7b518ce0c4eabcb150e112a1
Change-Id: Ie1f2f47097e84aa909cf2a6c937a017c44e32d47