[roll] Roll fuchsia [nelson][spi] Fix chip select GPIO name
The Amlogic SPI driver expects each GPIO fragment name to be in the form
"gpio-cs-n", where n is the chip index. The Nelson board driver was not
updated to use this naming scheme, so this change fixes it and adds a
device-enumeration test entry.

Test: runtests -n device-enumeration-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/467474
Original-Revision: 72606c42f45806fa109df7c914c3d15f6b88713d
GitOrigin-RevId: 644a504daa0214a83461eb11d0d4bd6bece9f2f1
Change-Id: I9f70997b7b139247b71dee1a354cfa8626f3899f