[roll] Roll fuchsia [ffx] Timeout identify_host call in OvernetPeer handler.
In cases where a target reboots after the daemon has opened an RCS
connection to it, the call to identify_host in
Target::from_rcs_connection can sometimes hang indefinitely. This
appears to be a cause of issue 66905.

Adding a timeout to this call prevents the hang. I'm so far unable to
repro issue 66905 with this change.

Original-Bug: 66905
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/467203
Original-Revision: f05fcb146ea070b69b7e6fe8fde5f1451a34dcee
GitOrigin-RevId: 557a680e344279830533fe48e369d7578b03681b
Change-Id: Ic7d2cdb730d8392abdb766a9f657af30c5e16bef