[roll] Roll fuchsia [storage]: Ftl Integration Test.
Fixed test that was not building anymore due to a previous CL which
added a ftl root node to the inspect hierarchy.

Added build only integration test, so it doesnt break again. This is
achieved by setting the environment as an empty list for the package.

TEST=CQ

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/467205
Original-Revision: 1d852a71c9608267048d1c48422fd99cc8bdba3d
GitOrigin-RevId: 064e6f3b094201b343f95e903c193c4d5073a539
Change-Id: I66eea6e0b6640261dd5307a82af1c4094ad09671