[roll] Roll fuchsia [ffx] Clean up 'ffx component' and add additional help text
The goal is to have a comprehensive help and examples so that the user
is able to understand the commands and operations.

Also provides some guidance for devs who are contributing plugins to
follow a similar structure and language for the help text.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/466856
Original-Revision: c0a43a1fe429b8ef206e18a886d2c8f9efc0b075
GitOrigin-RevId: 2dcd9c7ce0467cf2e1d95034aa4f3cfc5d19b578
Change-Id: I0500a82e2c78c7bdd9406c837ec2715cbe8fd3ac