[roll] Roll fuchsia [time][e2e] Use setupAll for SL4F setup
On occasion this test times out, and the retry fails
due to a "client closed" error. The SL4F client is
closed when teardown is called. This updates the test
to setup and teardown SL4F once for the whole set of
tests in an attempt to reduce instances of the error.
Since SL4F listens on a hardcoded port, the test cases
already share an instance of SL4F if run concurrently.

Original-Bug: 67173

Multiply: host_x64/time_sync_test

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/467854
Original-Revision: 6c64dbb8170dacfa21118c69935737aef64f85ab
GitOrigin-RevId: e863aafa9c30d603754a841b418f184176f4a07f
Change-Id: I3550beeb694c7607ba5910e944bf60545068fc71