[roll] Roll fuchsia [testsharder] Prefer less runs for tests listed in multiple modifiers.
A test may appear in an affected tests file which will cause it to be
run up to a default max number of times or as many times as will fit in
the target duration according to test duration data. However, this may
cause some issues in certain cases like go/bbid/8860185618054452560
where the affected test had many test cases and produced a really large
amount of logs that the builder failed due to running out of memory. In
that case, we should be able to explicitly set a smaller number of runs
through the MULTIPLY feature. However, that causes two multipliers to be
created for the same test (one from MULTIPLY and one from the affected
tests file). The current testsharder takes the number of runs associated
with the last multiplier in the list (the one from the affected tests).
This change will make it take the min number of runs from all
multiplers for the same test.

MULTIPLY: testsharder: 500
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/465054
Original-Revision: 9e3ac798e9d484d99b5b6bfbba58890fdab644b8
GitOrigin-RevId: 1c6e9b05b9abcae807588e3f80bf05c1269a676f
Change-Id: Ibe5a426d7bbbd9ac12df2ba7737060cad0ab4749