[roll] Roll fuchsia [fx ota] wait for commit post-reboot
This change modifies `fx ota` so that post-reboot, it waits for the
update to commit. If the commit hangs for more than 30 seconds, a
message will print to recommend reverting the update with `update
revert`.

Test: fx self-test subcommands/fx_ota_test
Test: Manually observed it print. Also, added a timer to stall the
commit and observed the "hang for 30 seconds" printline.
Original-Fixed: 64604

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/467197
Original-Revision: 5117e0ef0489b857ce5eadaecf4e26d497d1ef38
GitOrigin-RevId: dcf9095aa0bc2bd50036a4b2c119b75abb541743
Change-Id: I19f3cc1014bd69fbea8f3a066e6580127a3d967b