[roll] Roll fuchsia [kernel][kasan] Quarantine pmm frees to maximize reuse distance
Kernel Address Sanitizer is a kernel memory error detector that
detects certain use-out-of-bounds and use-after-free errors via
compiler instrumentation and sideband information.

In order to detect use-after-free errors, address sanitizer would
prefer that the time between a buffer is freed and the time it
is reused is maximized; in the kernel heap, we "quarantine" freed
buffers for a short time before releasing them to the heap free
lists.

This CR modifies the PMM to insert freed pages at the tail, rather
than the head of the free list (the PMM allocates from the head of
the free list) in address sanitizer builds. Inserting at the tail
maximizes the time between free()ing a page and subsequent
reallocations of that page without additional memory use.

In the future, this feature may be desirable at runtime while the
PMM free-fill checker is active; defer that for now, as the
performance characteristics are not yet understood.

Original-Bug: 30033
Pair-Programmed-With: mvanotti@google.com
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/467598
Original-Revision: fe9fe699db327a8b56eb58d8658dffac08a49a79
GitOrigin-RevId: 02fdb6418a9d9f46e5c3e0279186087e084f79b5
Change-Id: I2d61f8311a253e15e099611204d5f9306f040040