[roll] Roll infra/recipes [status_check] Use everywhere
Replace all uses of `api.test()` with `api.status_check.test()` to
comply with the new style guide (see
http://go/fxi/recipes#status-check-test).

This was a mostly a mechanical change, adding the expected status to all
test cases where the recipe currently fails. The only non-mechanical
change was in the cipd_with_dependencies_roller recipe, where a test
that was intended to pass was actually failing because it used dry-run
mock roll data for a live-run test. The solution was to switch the test
to dry-run mode.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/468115
Original-Revision: 021483960554b1a32df77912e4d767f62adc0505
GitOrigin-RevId: b72a1c1d9c3d292e5a1b261b93dec23e7ea2f59d
Change-Id: I0eb35a3c9ebb5b79a365f2910e8e7abc41213168