[roll] Roll fuchsia [build] Fix thinlto bringup build.
This fixes two issues that prevent the bringup.x64-thinlto build from
succeeding, as described in the associated bug entry:

  1) All source references to libzircon.so should point to the same
     binary when the "thinlto" variant is enabled.

  2) libzircon.so should never be installed into Fuchsia packages or
     as ZBI bootfs entries, since it is baked in the kernel.

Additionally, this fixes another problem that the fixes above reveal,
which is the fact that the libvulkan_goldfish.so library was not in
the appropriate variant toolchain, and thus added a dependency to
x64-shared/libfdio.so, which conflicted with
x64-thinlto-shared/libfdio.so at packaging time.

This is fixed by introducing a new variant_shared_library_redirect()
template and using it to build the library directly (i.e. without it
being a dependency of executable() or loadable_module()).

Another issue is the fact that libvulkan_goldfish.so fails to build
with the Asan variant, because it forces C++ RTTI. This is fixed
by excluding instrumented variants for this target.

Note: This does not fix other issues with the thinlto build, for
examples, see https://fxbug.dev/67097 or https://fxbug.dev/67099.

Original-Bug: 67096
Original-Bug: 67103

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/467134
Original-Revision: 74d38e7537819f3040ca47a118d7cb6576a604ac
GitOrigin-RevId: c988ac269b7b9b78c468b4d92685debbcede4809
Change-Id: I1c9ffdaec3c9b9ae85ca03a32d783da197ee2b61