[roll] Roll fuchsia [bt][common] Fix move semantics of AdvertisingData
The move semantics of AdvertisingData were unclear. A test assumed
that a moved-from AD would be in a freshly-initialized state, but this
failed to be true in a number of common states (e.g. whenever any of
the optional fields of the AD were set). Change and document the move
ctor to be user defined s.t. a moved-from object is reset to a freshly-
initialized state.

Original-Fixed: 66979
Test: `$ fx test bt-host-unittests`, added to
      GAP_AdvertisingDataTest.Move.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/466165
Original-Revision: 29b3982cf38cc3375f535c37202b96707f61f273
GitOrigin-RevId: 5727aa58dbc8a818ff3e52a25b9384a85bfc2eea
Change-Id: I4c75f01b2880c1eaa7ef8fe2ac52f1a1286dec88