[roll] Roll fuchsia [debugger] Handle deleting a target properly.
Previously when deleting the current target, it would pick the first
one. But this was done before the old one was deleted. The result was
that ConsoleContext would pick the same target that's about to be
deleted if the first one is deleted.

This reorders the logic to avoid this case.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/468236
Original-Revision: c351f8030f205a49b7b4f991d7d6d4a1ab98fc17
GitOrigin-RevId: 94d1a3d2720cd39d12c042bcbfe849f975ff2889
Change-Id: I97cce3020ef770300dbee3567a3394268d1755f1