[roll] Roll fuchsia [flake][time] Increase time e2e test timeout
While time synchronization itself is expected to complete within a
minute in most cases, on emulators the timeout can still be exceeded if
the device is busy. This seems to occur because the requires sl4f to be
started via ssh and to respond to http requests, which do not complete
in time.

Original-Bug: 67173

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/468240
Original-Revision: 401664dcddaabbaeb7feccdd7c40cd8adff5eeb3
GitOrigin-RevId: 891b88c60d05478c2149556cc33413c62189c483
Change-Id: Ia88904027aecf3682c18b8b48a58b0a585f0a127