[roll] Roll fuchsia [fx] Better message when goma is not started
If goma is enabled but not started, 'fx build' will present a generic
error message that instructs the user to run 'goma_ctl.py' instead of
the 'fx goma' wrapper. There is no way to make that message better
because it is generic for all goma customers.

This CL checks if goma is enabled in args.gn and then prints an
fx-specific error message if Goma is not started. This happens only
after a failed Ninja call, so that successful builds don't pay the price
to check goma status.

Original-Fixed: 67059
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/466694
Original-Revision: 0223b05afa5095723f246b81430d2329a332fe7d
GitOrigin-RevId: eb91d009a6052bf3e777fa5ee28f0efcee1b82a3
Change-Id: I75ade9b2f56ffdb928ad9e6f30d34584bf40ec6d