[roll] Roll fuchsia [fvm] Define board/product level dimension args
Allows callers of generate_fvm() to specify blobfs/minfs dimensions on a
product level or a board level.

If product-level arguments are passed, they are preferred over
board-level arguments. This is sufficient for our current use cases, but
in the future we may want to use board-level as a minimum setting since
the board would presumably determine the physical limits on the
partition dimensions (which the product could further restrict if
desired).

This change is a NOP for the current build, but allows a soft transition
onto these new board/product dimension arguments. After all clients are
onto the new arguments, we can clean up the old arguments.

Eventually we will remove the different configuration for sparse blobfs
instances, and use a product-level override to accomplish the same goal.

Original-Bug: 67015
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/465989
Original-Revision: 3f7183be92912daed083cfaa9156eb724271d93d
GitOrigin-RevId: f1d67940f3fa61cce86912954b5efb572f960977
Change-Id: Id52b5f455041f302d26e14fc6af3eefb4d8f1717