[roll] Roll fuchsia [kernel][debuglog] Avoid shutdown race
It is possible, albeit unlikely, for multiple threads to attempt to
shutdown the debuglog. One way this could occur could be for the system
to OOM and at the same time the critical job to be killed.

The solution here is to detect multiple shutdown requests, and have the
second request simply wait its maximum possible time.

Original-Bug: 57008

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/468154
Original-Revision: c83d6613288351ddee40d49b10b200a0c41759a8
GitOrigin-RevId: 09a21aae6cf9f5027fa6fd7754617f65b0942810
Change-Id: I63bfd7ff41190864bf857c106597061558072d2e