[roll] Roll fuchsia [display][hdmi][i915] Add audio support
We pass ELD (EDID like data as defined in the Intel HDA spec) from
the display controller (gets it from the EDID) to the display
implementationdriver (i915). For this we:

1. Create a small ELD library to be used by HDMI display drivers to
be shared with any future HDA audio HDMI drivers.
2. Add a SetEld banjo method independent from ApplyConfiguration only
to be called when the ELDs change.
3. Extract ELDs from EDIDs (right after EDID reads) and call SetEld
so display implementation drivers can make use of them (we use it
in the i915).
4. Update the fake display driver to provide canned EDIDs via I2C.
5. Trigger HDA ELD functionality in the i915 driver to be used by an
audio HDMI driver (in an upcoming CL).

Test: display-core-unittests
Test: Manual HDMI audio check with upcoming HDA audio CL.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/467857
Original-Revision: 50e263ebdb41a5ecb0943873eee1dbd10ca1074f
GitOrigin-RevId: 1c6434a2de69ae3fb34cb5769e161cd6210215e7
Change-Id: I0a8bb3bf60e4cbaa539d0867999ef86cab477c96