[roll] Roll fuchsia [kernel][arch] Refactor CPUID corpus
Current use of the CPUID corpus is confined to a single test suite and
relies on file I/O and JSON parsing, making it unusable for kernel
tests. It will be convenient to use the CPUID values for specific
microprocessors in kernel tests, as is already done with cpu_id::CpuId
today. Accordingly, this change rethinks the stored format and the means
of loading that data into a FakeCpuidIo object in order to make this
more kernel-friendly.

Following up on a past suggestion of mcgrathr@'s, we can leverage the
preprocessor for file I/O. The format is updated to consist of lines
that look like
```
DEFINE_CPUID_VALUES(<leaf>, <subleaf>, <eax>, <ebx>, <ecx>, <edx>)
```
so that one may `#include` after defining a DEFINE_CPUID_VALUES() macro:
e.g.,
```
\#define DEFINE_CPUID_VALUES(leaf, subleaf, eax, ebx, ecx, edx) ...
\#include "data/cpuid/my-microprocessor.inc"
\#undef DEFINE_CPUID_VALUES
```

This removes the need for runtime file I/O and JSON support. Further, we
parametrize a FakeCpuidIo constructor on an enum value coinciding with
one of these files, moving the management of the corpus into
//zircon/kernel/lib/arch/testing.

The corpus tests continue to pass and are made to use the new
FakeCpuidIo constructor.

See https://fuchsia-review.googlesource.com/c/fuchsia/+/465895 for an
example of what it looks like now to extend the corpus - and
https://fuchsia-review.googlesource.com/c/fuchsia/+/465898 for an
example of use within the kernel.

Original-Bug: 61093
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/465894
Original-Revision: 126ab1517b3e3b470ab24d92c463cf1ed7ac69f3
GitOrigin-RevId: e2bc8a87da92fca36126a3720a6e46d5c360058c
Change-Id: I0a87232bb4eb701122e2e56ba13df92fb12840e3