[roll] Roll fuchsia [memalloc] Add RangeStorage class for allocating storage space.
The current memalloc::Allocator API requires a fbl::Span<Range> of
pre-allocated storage to be passed in to the constructor. This storage
is used as the internal range nodes of the allocator.

However, the Range type is not POD, meaning that global/static
allocations need to call the constructor when we first use the storage,
and emit calls to the destructor on program termination:

  static memalloc::Range storage[100];  // non-POD
  memalloc::Allocator allocator(storage);

This isn't ideal, and causes link errors in minimal environments such as
physboot which don't have the appropriate symbols defined to support
global destruction.

Instead, we introduce a new RangeStorage class that can be treated like
a POD class and avoids these issues, and leave explicit
construction/destruction up to the memalloc::Allocator class:

  static memalloc::RangeStorage storage[100];
  memalloc::Allocator allocator(storage);

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/465096
Original-Revision: fdc9871ca0338c371b87f93416a778d34d2a60a1
GitOrigin-RevId: 834d31700b281bf6172056abd01bd8e51deff6f5
Change-Id: I2569c6946c2e294a2f01337d6f37ec57bfa39b54