[roll] Roll fuchsia [build] Modernize build support for grand unified binaries
Unfortunately also piled on a bunch of entangled cleanups. :/
Everywhere that the Go grand_unified_binary and the Rust
grand_swd_binary went and I had to replace with a resource-based target
needed to be converted to modern package & component templates or else
chaos and inconsistency would have ensued.

Breaking this change to smaller changes would have required multiple
soft transitions because size_checker gets confused about who to blame
grand_unified_binary and associated libs in the interim state.

The changes however are nearly no-op in terms of the produced image. The
only things that changed is some cmx/cml files referenced a
program.binary with a renamed name (e.g. Rust binary with dashes instead
of underscores) that was inconsistent with what the underlying
executable generated, so I fixed that.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/466814
Original-Revision: 4b330447349193f27f9f47afe217013730b942e3
GitOrigin-RevId: 551344c361920fd9dfd120c452701a531992e026
Change-Id: Ic5727bcb871ace98208680a93eb829fd83a405da